requirements.txt
httpx==0.27.2
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
sqlalchemy>=2.0.40
jinja2==3.1.4
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
